TOP-2394 - add submenu for product
