+++ src/lib/components/LinksMenu.module.scss
@@ -96,3 +96,7 @@
     }
   }
 }
+
+.menuItemWrap {
+  position: relative;
+}
