--- src/lib/components/hover-menu/HoverMenu.module.scss
@@ -1,143 +1,0 @@
-@import 'lib/styles/mixins.scss';
-@import 'lib/styles/fonts.scss';
-@import 'lib/styles/colors.scss';
-
-$transitionInDelay: 0.2s;
-
-.hoverMenuWrap {
-  position: absolute;
-  top: 100%;
-  left: 0;
-  right: 0;
-
-  background: #fff;
-  box-shadow: 0px 6px 24px rgba(0, 0, 0, 0.1);
-  border-radius: 0px 0px 8px 8px;
-
-  font-family: $roboto;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 22px;
-  color: #2A2A2A;
-
-  visibility: hidden;
-  opacity: 0;
-  z-index: 2;
-  transform: translateY(5px);
-  pointer-events: none;
-
-  // transition out
-  transition: opacity 0.2s ease-out 0.15s,
-    visibility 0s linear .35s,
-    transform 0.15s ease-out 0.15s;
-
-  :global(.has-menu:not(.active)):hover ~ &, &:hover {
-    opacity: 1;
-    visibility: visible;
-    z-index: 19;
-    transform: translateY(0);
-    pointer-events: auto;
-
-    // transition in
-    transition: opacity 0.2s ease-in calc(0.15s + $transitionInDelay),
-      visibility 0s linear 0.35s,
-      transform 0.15s ease-in calc(0.15s + $transitionInDelay);
-  }
-}
-
-.hoverMenuInner {
-  position: relative;
-  padding: 32px;
-  @include maxViewWidth;
-  margin: 0 auto;
-}
-
-.mainDesc {
-  padding: 0;
-  margin: 0;
-  font-size: 14px;
-  line-height: 22px;
-}
-
-.menuSections {
-  display: flex;
-  align-items: flex-start;
-  gap: 48px;
-  margin-top: 24px;
-}
-
-.menuSectionItem {
-  flex: 1 1 0;
-  font-size: 12px;
-  line-height: 18px;
-}
-
-.ctasMenuSectionItem {
-  flex: 1 1 0;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  align-items: flex-start;
-}
-
-a.menuSectionHeading {
-  all: unset;
-  display: block;
-  color: color(hoverMenuLink);
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 700;
-  cursor: pointer;
-  &:hover {
-    color: color(hoverMenuLink, hover);
-  }
-  &:global(.active), &:active {
-    color: color(hoverMenuLink, active);
-  }
-}
-
-.menuSectionDesc {
-  all: unset;
-}
-
-.menuSectionItemChilds {
-  > li {
-    margin-top: 12px;
-  }
-}
-
-a.menuSectionChild {
-  all: unset;
-  font-weight: 500;
-  color: color(hoverMenuLink);
-  cursor: pointer;
-  &:hover {
-    color: color(hoverMenuLink, hover);
-    font-weight: 700;
-  }
-  &:global(.active), &:active {
-    color: color(hoverMenuLink, active);
-    font-weight: 700;
-  }
-}
-
-a.ctaButton {
-  font-family: $roboto;
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 24px;
-  letter-spacing: 0.008em;
-
-  background: #137D60;
-  color: #fff;
-  border: 1px solid #137D60;
-  border-radius: 24px;
-
-  padding: 0 16px;
-  transition: background 0.15s ease-in-out;
-
-  &:hover {
-    background: #219174;
-  }
-}
