--- src/lib/components/hover-menu/index.ts
@@ -1,1 +1,0 @@
-export { default as HoverMenu } from './HoverMenu.svelte'
