+++ src/lib/config/nav-menu/main-navigation.config.ts
@@ -3,7 +3,13 @@ import type { NavMenuItem } from "../../functions/nav-menu-item.model";
 import { allNavItems } from "./all-nav-items.config";
 
 export const mainNavigationItems: NavMenuItem[] = [
-  allNavItems.product,
+  {
+    ...allNavItems.product,
+    children: [
+      allNavItems.bugHunt,
+      allNavItems.platform,
+    ],
+  },
   allNavItems.pricing,
   allNavItems.demo,
   allNavItems.talent,
